Add the Hidex USB workaround for old misconfigured-firmware devices
Just copied the change from the 4.19-hidex branch to this new branch,
because the upstream beagleboard/linux/4.19 branch is funky and seems
to be re-applying a set of patches on top of newer upstream Linux
kernel, instead of doing merges. And thus just merging that
beagleboard/linux/4.19 branch does not apply cleanly.